fixed a typo in the /= operator
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@4394 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ lib/QoreDivideEqualsOperatorNode.cpp
@@ -45,7 +45,7 @@ AbstractQoreNode *QoreDivideEqualsOperatorNode::evalImpl(ExceptionSink *xsink) c
 
    // is either side a float?
    if ((res && res->getType() == NT_FLOAT) || v.get_type() == NT_FLOAT) {
-      float val = res ? res->getAsFloat() : 0.0;
+      double val = res ? res->getAsFloat() : 0.0;
       if (val == 0.0) {
 	 xsink->raiseException("DIVISION-BY-ZERO", "division by zero in floating-point expression");
 	 return 0;
